% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biase.R
\docType{data}
\name{cell_type_biase}
\alias{cell_type_biase}
\title{Biase Data Cell Type}
\format{An R.Data object containing a vector with the cell type of each of the
cells in the Biase Data.}
\usage{
cell_type_biase
}
\description{
The cell type of each of the cells in the Biase data.
}
\keyword{datasets}
