% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biase.R
\docType{data}
\name{condition_biase}
\alias{condition_biase}
\title{Biase Data Conditions}
\format{An R.Data object containing a vector with the conditon of the 49 cells
in the Biase data.}
\usage{
condition_biase
}
\description{
The condition for each sample in the Biase data. To be used when splitting
the data to demonstrate SparseDC.
}
\keyword{datasets}
