% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biase.R
\docType{data}
\name{data_biase}
\alias{data_biase}
\title{Biase Data}
\format{An R.Data object storing FPKM gene expression
measurements for each of the samples.}
\usage{
data_biase
}
\description{
This dataset was created by Biase et al. to study cell fat inclination in
mouse embryos. It contains FPKM gene expression measurements for 49 cells
and 16,514 genes. There are three cell types in the dataset, zygote,
two-cell embryo, and four-cell embryo cells.
}
\keyword{datasets}
