% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsemdc_gap.R
\name{generate_uni_dat}
\alias{generate_uni_dat}
\title{Uniform data generator
For use with the gap statistic. Generates datasets drawn from the reference
distribution where each reference feature is generated uniformly over the
range of observed values for that feature.}
\usage{
generate_uni_dat(data)
}
\arguments{
\item{data}{A dataset with rows as features and columns as samples.}
}
\value{
A dataset drawn from the reference distribution for use internally
with the gap statistic.
}
\description{
Uniform data generator
For use with the gap statistic. Generates datasets drawn from the reference
distribution where each reference feature is generated uniformly over the
range of observed values for that feature.
}
