% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_calc.R
\name{mu_calc}
\alias{mu_calc}
\title{mu Calculator}
\usage{
mu_calc(d, v, EQ, dim, x, nk, p1, p2, delta)
}
\arguments{
\item{d}{- The current condition.}

\item{v}{- Relationship matrix, describing relationship between d-1 and d.}

\item{EQ}{- Equality matrix specifying the number of following terms to
which mu_d is equal.}

\item{dim}{Total number of conditions, D.}

\item{x}{- Matrix of mean values.}

\item{nk}{- Vector with the number of samples in each dimension for this
cluster.}

\item{p1}{- Penalties on mean size.}

\item{p2}{- Penalties on mean difference.}

\item{delta}{Small term to ensure existance of solution.}
}
\value{
A list containing two vectors containing the calculated values of
mu_{d} | mu_{d} < mu_{d-1} and mu_{d} | mu_{d} > mu_{d-1} respecitively.
}
\description{
This function handles the calculations of the Mu Solver. This function
 runs inside \code{sparse_mdc}.
}
