% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen_calculator.R
\name{pen_calculator}
\alias{pen_calculator}
\title{Penalty calculator}
\usage{
pen_calculator(lambda1, lambda2, nk, delta)
}
\arguments{
\item{lambda1}{Calculated penalty parameter for mean size.}

\item{lambda2}{Calculated penalty parameter for mean difference.}

\item{nk}{Vector containing the number of samples in each dimension.}

\item{delta}{Small term to ensure existance of solution, default is
0.0000001.}
}
\value{
a list with two vectors containing the penalty terms for mean size
and mean difference respectively.
}
\description{
Calcualtes the penalty terms for penalizing mean size and mean difference.
This function runs inside \code{sparse_mdc}.
}
