% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_func.R
\name{S_func}
\alias{S_func}
\title{The soft thresholding operator}
\usage{
S_func(x, a)
}
\arguments{
\item{x}{The data value}

\item{a}{The lambda value}
}
\value{
The solution to the soft thresholding operator.
}
\description{
Function to solve the soft thresholding problem
}
