% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calc.R
\name{score_calc}
\alias{score_calc}
\title{Score calculator}
\usage{
score_calc(pdat, clusters, mu, lambda1, lambda2, nclust, delta, dim)
}
\arguments{
\item{pdat}{list with D entries, each entry contains data d, p * n matrix.
This data should be centered and log-transformed.}

\item{clusters}{List containig cluster assignments for each dimension
as entries.}

\item{mu}{list with D entries, each entry contains centers for data d,
p*k matrix.}

\item{lambda1}{Penalty parameter for mean size.}

\item{lambda2}{Penalty parameter for mean difference.}

\item{nclust}{Number of clusters in the data.}

\item{delta}{Small term to ensure existance of solution, default is
0.0000001.}

\item{dim}{Total number of conditions, D.}
}
\value{
The caluculated score.
}
\description{
Calculates the score for each combination of cluster assignments
and center values. This function runs inside \code{sparse_mdc}.
}
