% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_c.R
\name{update_c}
\alias{update_c}
\title{update_c}
\usage{
update_c(mu, pdat, nclust, dim, lambda1, lambda2, delta)
}
\arguments{
\item{mu}{list with D entries, each entry contains centers for data d,
p*k matrix.}

\item{pdat}{list with D entries, each entry contains data d, p * n matrix.
This data should be centered and log-transformed.}

\item{nclust}{Total number of clusters.}

\item{dim}{Total number of conditions, D.}

\item{lambda1}{Calculated penalty parameter for mean size.}

\item{lambda2}{Calculated penalty parameter for mean difference.}

\item{delta}{Small term to ensure existance of solution.}
}
\value{
A list with D entries containing cluster assignments for each
sample.
}
\description{
Assigns samples to clusters. This function runs inside
 \code{sparse_mdc}.
}
