% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mu.R
\name{update_mu}
\alias{update_mu}
\title{update_mu}
\usage{
update_mu(clusters, pdat, nclust, dim, lambda1, lambda2, ngenes, delta)
}
\arguments{
\item{clusters}{List containig cluster assignments for each dimension
as entries.}

\item{pdat}{list with D entries, each entry contains data d, p * n matrix.
This data should be centered and log-transformed.}

\item{nclust}{Number of clusters in the data.}

\item{dim}{Total number of conditions, D.}

\item{lambda1}{Calculated penalty parameter for mean size.}

\item{lambda2}{Calculated penalty parameter for mean difference.}

\item{ngenes}{The number of genes in the data.}

\item{delta}{Small term to ensure existance of solution.}
}
\value{
A list containing the center values for the clusters in each dimensions
 as entries.
}
\description{
Update the mean/center values for each cluster and dimension.
 This function runs inside \code{sparse_mdc}.
}
